+++ src/androidMain/kotlin/KSLoggerDefaultPlatformLoggerLambda.kt
@@ -2,10 +2,10 @@ package dev.inmo.kslog.common
 
 import android.util.Log
 
-
-internal actual val defaultLogging: (level: LogLevel, tag: String, message: Any, throwable: Throwable?) -> Unit = { l, t, m, e ->
+actual var KSLoggerDefaultPlatformLoggerLambda: (level: LogLevel, tag: String, message: Any, throwable: Throwable?) -> Unit = { l, t, m, e ->
     val messageString = m.toString()
     when(l) {
+        LogLevel.TRACE -> Log.d(t, messageString, e)
         LogLevel.DEBUG -> Log.d(t, messageString, e)
         LogLevel.VERBOSE -> Log.v(t, messageString, e)
         LogLevel.INFO -> Log.i(t, messageString, e)