+++ src/commonMain/kotlin/KSLog.kt
@@ -5,6 +5,7 @@ import dev.inmo.kslog.common.utils.plus
 
 
 enum class LogLevel {
+    TRACE,
     DEBUG,
     VERBOSE,
     INFO,
@@ -13,15 +14,38 @@ enum class LogLevel {
     ASSERT,
 }
 
+/**
+ * Base interface for any logger
+ *
+ * @see default
+ *
+ */
 interface KSLog {
+    /**
+     * The only one function required to realization in any inheritor. All other [performLog] functions
+     * will call this one by default
+     */
     fun performLog(level: LogLevel, tag: String?, message: Any, throwable: Throwable?)
+
+    /**
+     * Calls default [performLog] with `tag` == `null`
+     */
     fun performLog(level: LogLevel, message: Any, throwable: Throwable?) = performLog(level, null, message, throwable)
+    /**
+     * Calls default [performLog] with `message` built using [messageBuilder]. This method supposed to be overriden in
+     * case when logger supports lazy-loaded messages (like [dev.inmo.kslog.common.filter.FilterKSLog])
+     */
     fun performLog(
         level: LogLevel,
         tag: String?,
         throwable: Throwable?,
         messageBuilder: () -> Any
     ) = performLog(level, tag, messageBuilder(), throwable)
+
+    /**
+     * Suspendable variant of [performLog] with [messageBuilder]. Uses default [performLog] with `message` built using
+     * [messageBuilder] by default
+     */
     suspend fun performLogS(
         level: LogLevel,
         tag: String?,
@@ -65,15 +89,20 @@ interface KSLog {
 
 operator fun KSLog.invoke(performLogCallback: (level: LogLevel, tag: String?, message: Any, throwable: Throwable?) -> Unit) = CallbackKSLog(performLogCallback)
 
-internal expect val defaultLogging: (level: LogLevel, tag: String, message: Any, throwable: Throwable?) -> Unit
 internal val printlnLogging: (level: LogLevel, tag: String, message: Any, throwable: Throwable?) -> Unit = { l, t, m, e ->
     println(defaultMessageFormatter(l, t, m, e))
 }
 
+/**
+ * Simple builder for [DefaultKSLog] logger based on [defaultTag]
+ */
 fun KSLog(
     defaultTag: String,
 ): KSLog = DefaultKSLog(defaultTag)
 
+/**
+ * Simple builder for [DefaultKSLog] logger based on [defaultTag] and [messageFormatter]
+ */
 fun KSLog(
     defaultTag: String,
     messageFormatter: MessageFormatter
@@ -92,17 +121,25 @@ fun KSLog(
     messageFormatter
 ).filtered(filter)
 
+/**
+ * Building logger using [KSLog] builder based on [defaultTag] and [messageFormatter]. This logger will also filter
+ * incoming levels: only levels passed in [levels] param will be logged
+ */
 fun KSLog(
     defaultTag: String,
     levels: Iterable<LogLevel>,
     messageFormatter: MessageFormatter = defaultMessageFormatter
 ): KSLog {
-    val levels = levels.toSet()
+    val levelsSet = levels.toSet()
     return KSLog (defaultTag, messageFormatter).filtered { l, _, _ ->
-        l in levels
+        l in levelsSet
     }
 }
 
+/**
+ * Building logger using [KSLog] builder based on [defaultTag] and [messageFormatter]. This logger will also filter
+ * incoming levels: only levels passed in [firstLevel], [secondLevel] and [otherLevels] param will be logged
+ */
 fun KSLog(
     defaultTag: String,
     firstLevel: LogLevel,
@@ -111,6 +148,10 @@ fun KSLog(
     messageFormatter: MessageFormatter = defaultMessageFormatter,
 ): KSLog = KSLog(defaultTag, setOf(firstLevel, secondLevel, *otherLevels), messageFormatter)
 
+/**
+ * Building logger using [KSLog] builder based on [defaultTag] and [messageFormatter]. This logger will also filter
+ * incoming levels: only levels above [minLoggingLevel] will be logged
+ */
 fun KSLog(
     defaultTag: String,
     minLoggingLevel: LogLevel,
@@ -122,5 +163,13 @@ fun KSLog(
     minLoggingLevel.ordinal <= l.ordinal
 }
 
+/**
+ * Setting [KSLog.default] logger to [newDefault]
+ */
 fun setDefaultKSLog(newDefault: KSLog) { KSLog.default = newDefault }
-fun addDefaultKSLog(newDefault: KSLog) { KSLog.default = KSLog.default + newDefault }
+/**
+ * Setting [KSLog.default] logger to new [CallbackKSLog] using [plus] operation
+ */
+fun addDefaultKSLog(newDefault: KSLog) {
+    KSLog.default += newDefault
+}